// src/types/features.ts
export type Feature = {
  id: number;
  name: string;
  description: string;
  color: string;
  icon: string;
  is_active: number;
  created_at: string;
  updated_at: string;
};

export type badge = {
  id: number;
  name: string;
  description: string;
  color: string;
  icon: string;
  is_active: number;
  created_at: string;
  updated_at: string;
};